Add support for reading the 'ReferencedSegmentNumber' from the 'PerFrameFunctionalGroupsSequence' or the 'SharedFunctionalGroupsSequence'
The tag 'ReferencedSegmentNumber' can appear in either the 'PerFrameFunctionalGroupsSequence' or 'SharedFunctionalGroupsSequence' if it is shared for all the frames. This is stated in the DICOM format:
"Those Functional Groups that apply to all frames are included in the Shared Functional Groups Sequence (5200,9229). Functional Groups whose Attribute values may vary from frame to frame are included in the Per-frame Functional Groups Sequence (5200,9230)." A bit more detail in  http://dicom.nema.org/medical/dicom/current/output/chtml/part03/sect_C.7.6.16.html#sect_C.7.6.16.1.1 . Not sure, if the MultiClassReader should have a similar modification.