Add support for reading the 'ReferencedSegmentNumber' from the 'PerFr… (#12)
* Add support for reading the 'ReferencedSegmentNumber' from the 'PerFrameFunctionalGroupsSequence' or the 'SharedFunctionalGroupsSequence'

The tag 'ReferencedSegmentNumber' can appear in either the 'PerFrameFunctionalGroupsSequence' or 'SharedFunctionalGroupsSequence' if it is shared for all the frames. This is stated in the DICOM format:
"Those Functional Groups that apply to all frames are included in the Shared Functional Groups Sequence (5200,9229). Functional Groups whose Attribute values may vary from frame to frame are included in the Per-frame Functional Groups Sequence (5200,9230)." A bit more detail in  http://dicom.nema.org/medical/dicom/current/output/chtml/part03/sect_C.7.6.16.html#sect_C.7.6.16.1.1 . Not sure, if the MultiClassReader should have a similar modification.

* Bug fix: Change dcm to dataset

* Add changes proposed by razorx89

Added the changes that you suggested to the SegmentReader and MultiClassReader. The MultiClassReader Changes were not tested because I do not have SEG file with multiple classes

* Remove auxiliary lines of code added in the previous commit